# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WBlackjack language learning app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SongStatus = typeof SongStatus[keyof typeof SongStatus];


export const SongStatus = {
  new: 'new',
  active: 'active',
  done: 'done',
} as const;

export interface Song {
  id: number;
  artist: string;
  title: string;
  youtubeUrl: string;
  language: string;
  status: SongStatus;
  timesPlayed: number;
  dateAdded: string;
  /** @nullable */
  lastPlayed?: string | null;
  hasLyrics?: boolean;
  hasTimestamps?: boolean;
}

export interface SongInput {
  /** @minLength 1 */
  artist: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  youtubeUrl: string;
  /** @minLength 1 */
  language: string;
}

export type SongUpdateStatus = typeof SongUpdateStatus[keyof typeof SongUpdateStatus];


export const SongUpdateStatus = {
  new: 'new',
  active: 'active',
  done: 'done',
} as const;

export interface SongUpdate {
  status?: SongUpdateStatus;
  timesPlayed?: number;
  /** @nullable */
  lastPlayed?: string | null;
}

export type SongStatsByStatus = {
  new: number;
  active: number;
  done: number;
};

export type SongStatsByLanguageItem = {
  language: string;
  count: number;
};

export interface SongStats {
  total: number;
  byStatus: SongStatsByStatus;
  byLanguage: SongStatsByLanguageItem[];
}

export interface LyricLine {
  id: number;
  songId: number;
  lineIndex: number;
  original: string;
  translation: string;
  /** @nullable */
  distractor1?: string | null;
  /** @nullable */
  distractor2?: string | null;
  /** @nullable */
  distractor3?: string | null;
  /** @nullable */
  distractor4?: string | null;
  /** @nullable */
  timestampMs?: number | null;
}

export interface LyricLineInput {
  lineIndex: number;
  original: string;
  translation: string;
  /** @nullable */
  distractor1?: string | null;
  /** @nullable */
  distractor2?: string | null;
  /** @nullable */
  distractor3?: string | null;
  /** @nullable */
  distractor4?: string | null;
}

export interface LyricsInput {
  lines: LyricLineInput[];
}

export interface TimestampEntry {
  lineIndex: number;
  timestampMs: number;
}

export interface TimestampsInput {
  timestamps: TimestampEntry[];
}

export type ListSongsParams = {
language?: string;
artist?: string;
status?: ListSongsStatus;
sort?: ListSongsSort;
search?: string;
};

export type ListSongsStatus = typeof ListSongsStatus[keyof typeof ListSongsStatus];


export const ListSongsStatus = {
  new: 'new',
  active: 'active',
  done: 'done',
} as const;

export type ListSongsSort = typeof ListSongsSort[keyof typeof ListSongsSort];


export const ListSongsSort = {
  date_added_desc: 'date_added_desc',
  date_added_asc: 'date_added_asc',
  last_played_desc: 'last_played_desc',
  last_played_asc: 'last_played_asc',
  title_asc: 'title_asc',
  title_desc: 'title_desc',
} as const;

export type ListArtistsParams = {
q?: string;
};

export type ListLanguagesParams = {
q?: string;
};