# lib/api-zod/src/generated/types/lyricLineInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WBlackjack language learning app API
 * OpenAPI spec version: 0.1.0
 */

export interface LyricLineInput {
  lineIndex: number;
  original: string;
  translation: string;
  /** @nullable */
  distractor1?: string | null;
  /** @nullable */
  distractor2?: string | null;
  /** @nullable */
  distractor3?: string | null;
  /** @nullable */
  distractor4?: string | null;
}