# lib/api-zod/src/generated/types/lyricsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WBlackjack language learning app API
 * OpenAPI spec version: 0.1.0
 */
import type { LyricLineInput } from './lyricLineInput';

export interface LyricsInput {
  lines: LyricLineInput[];
}