# lib/api-zod/src/generated/types/songInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WBlackjack language learning app API
 * OpenAPI spec version: 0.1.0
 */

export interface SongInput {
  /** @minLength 1 */
  artist: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  youtubeUrl: string;
  /** @minLength 1 */
  language: string;
}