# lib/api-zod/src/generated/types/songStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WBlackjack language learning app API
 * OpenAPI spec version: 0.1.0
 */
import type { SongStatsByLanguageItem } from './songStatsByLanguageItem';
import type { SongStatsByStatus } from './songStatsByStatus';

export interface SongStats {
  total: number;
  byStatus: SongStatsByStatus;
  byLanguage: SongStatsByLanguageItem[];
}