# lib/api-zod/src/generated/types/songUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WBlackjack language learning app API
 * OpenAPI spec version: 0.1.0
 */
import type { SongUpdateStatus } from './songUpdateStatus';

export interface SongUpdate {
  status?: SongUpdateStatus;
  timesPlayed?: number;
  /** @nullable */
  lastPlayed?: Date | null;
}