# lib/api-zod/src/generated/types/timestampEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WBlackjack language learning app API
 * OpenAPI spec version: 0.1.0
 */

export interface TimestampEntry {
  lineIndex: number;
  timestampMs: number;
}